# src/main/resources/apidocs/api-doc.json
{"openapi":"3.0.1","info":{"title":"Device Management API","description":"API for managing devices in the system","termsOfService":"https://www.api.com/terms-of-service","contact":{"name":"Support Team","url":"https://www.api.com/contact","email":"support@api.com"},"license":{"name":"Apache 2.0","url":"https://www.apache.org/licenses/LICENSE-2.0"},"version":"v1.0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"Device Management","description":"Operations for managing devices"}],"paths":{"/api/v1/devices/{id}":{"get":{"tags":["Device Management"],"summary":"Get device by ID","description":"Retrieve a device by its unique identifier","operationId":"getDeviceById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseDeviceDTO"}}}}}},"put":{"tags":["Device Management"],"summary":"Update an existing device","description":"Update an existing device's information","operationId":"updateDevice","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeviceDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseDeviceDTO"}}}}}},"delete":{"tags":["Device Management"],"summary":"Delete a device","description":"Remove a device from the system by its ID","operationId":"deleteDevice","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseVoid"}}}}}}},"/api/v1/devices":{"get":{"tags":["Device Management"],"summary":"List all devices","description":"Fetch a list of all registered devices","operationId":"listAllDevices","parameters":[{"name":"offset","in":"query","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"limit","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListDeviceDTO"}}}}}},"post":{"tags":["Device Management"],"summary":"Add a new device","description":"Create a new device with the provided information","operationId":"addDevice","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeviceDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseDeviceDTO"}}}}}}},"/api/v1/devices/search":{"get":{"tags":["Device Management"],"summary":"Search devices by brand","description":"Search and list devices by their brand","operationId":"searchByBrand","parameters":[{"name":"brand","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListDeviceDTO"}}}}}}}},"components":{"schemas":{"DeviceDTO":{"required":["brand","creationTime","name"],"type":"object","properties":{"id":{"type":"integer","description":"Unique identifier of the device","format":"int64","example":1},"name":{"type":"string","description":"Name of the device","example":"Smartphone X200"},"brand":{"type":"string","description":"Brand of the device","example":"XPhone"},"creationTime":{"type":"string","description":"Creation timestamp of the device","format":"date-time"}},"description":"Representation of a Device"},"ApiResponseDeviceDTO":{"type":"object","properties":{"result":{"$ref":"#/components/schemas/DeviceDTO"},"message":{"$ref":"#/components/schemas/ApiResponseMessage"},"pageable":{"$ref":"#/components/schemas/ApiResponsePageable"}}},"ApiResponseMessage":{"type":"object","properties":{"code":{"type":"string"},"description":{"type":"string"}}},"ApiResponsePageable":{"type":"object","properties":{"offset":{"type":"integer","format":"int32"},"limit":{"type":"integer","format":"int32"},"count":{"type":"integer","format":"int64"}}},"ApiResponseListDeviceDTO":{"type":"object","properties":{"result":{"type":"array","items":{"$ref":"#/components/schemas/DeviceDTO"}},"message":{"$ref":"#/components/schemas/ApiResponseMessage"},"pageable":{"$ref":"#/components/schemas/ApiResponsePageable"}}},"ApiResponseVoid":{"type":"object","properties":{"result":{"type":"object"},"message":{"$ref":"#/components/schemas/ApiResponseMessage"},"pageable":{"$ref":"#/components/schemas/ApiResponsePageable"}}}}}}